backend/requirements.txt
flask
flask-cors
google-genai
beautifulsoup4
requests
python-dotenv